requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.3
filelock==3.20.0
fsspec==2025.9.0
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community>=0.3.0
langchain-core==0.3.78
langchain-text-splitters==0.3.11
langsmith==0.4.33
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
ollama==0.5.1
orjson==3.11.3
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
protobuf==6.32.1
pydantic==2.12.0
pydantic_core==2.41.1
PyYAML==6.0.3
qdrant-client==1.15.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch>=2.9.0
tqdm==4.67.1
transformers==4.57.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
watchdog==4.0.1
zstandard==0.25.0
Flask==3.0.3
pyvis==0.3.2
fastmcp>=2.12.3

# Development dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0
isort==5.13.2
pandas>=2.0.0